requirements.txt
ollama==0.3.3
chromadb==0.4.24
sentence-transformers==2.2.2
langchain==0.1.20
langchain-community==0.0.38
pyyaml==6.0.1
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.4
numpy==1.26.2
tqdm==4.66.1
requests==2.31.0
